@@ -20,12 +20,13 @@
         }
     },
     "require": {
-        "php": ">=5.3.0",
-        "symfony/filesystem": "^2.3|^3|^4",
-        "symfony/process": "^2.3|^3|^4"
+        "php": ">=7.2.5",
+        "symfony/filesystem": "^4|^5",
+        "symfony/process": "^4|^5"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.5"
+        "phpunit/phpunit": "^8|^9",
+        "phpspec/prophecy-phpunit": "^2.0"
     },
     "extra": {
         "branch-alias": {
